Update: Refactored the File to contain a binary blob rather than a null-terminated string, and updated relevant callsites. Clean: Reduced warnings by adding StringFormat that defaults to safe string format call (snprintf_s) if available.

+++ Source/Engine/Resources/glTF/Accessor.cpp
@@ -199,10 +199,10 @@ namespace glTF
 
 		BufferView newView = origView;
 		newView.myByteOffset = 0; // baked in above
-		newView.myBuffer = aBuffers.size();
+		newView.myBuffer = static_cast<uint32_t>(aBuffers.size());
 
 		// repoint to new view
-		myBufferView = aBufferViews.size();
+		myBufferView = static_cast<uint32_t>(aBufferViews.size());
 		myByteOffset = 0; // baked in above
 
 		aBufferViews.push_back(std::move(newView));
